pyproject.toml
[project]
name = "research-template"
version = "0.1.0"
description = "A template for AI research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers>=4.48.3",
    "wandb>=0.19.6",
    "ruff>=0.9.5",
    "setuptools>=75.8.0",
    "flash-attn>=2.7.4.post1; sys_platform == 'linux'",
    "peft>=0.14.0",
    "pytest>=8.3.5",
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]
